+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 [_vNext_](https://github.com/sketch7/SignalR.Orleans/compare/1.0.0...1.1.0) (2019-X-X)
 
+
+## [3.1.0](https://github.com/sketch7/SignalR.Orleans/compare/3.0.0...3.1.0) (2021-12-02)
+
+### Features
+
+- **orleans:** Orleans 3.5.1 update
+- **client:** add log when connection not found
+- **orleans:** add log if stream fails to send on `ClientGrain`
+
 ## [3.0.0](https://github.com/sketch7/SignalR.Orleans/compare/2.0.0...3.0.0) (2021-04-29)
 
 ### Features